% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLV.R
\name{simlv}
\alias{simlv}
\title{Simulate a Lotka-Volterra model with noise.}
\usage{
simlv(n0, r, alphaij, tmax, mean_d = 0, sd_d = 0.1, ht = 1)
}
\arguments{
\item{n0}{The initial condition of the density of species 1 X num_species.}

\item{r}{The intrinsic growth rates of species 1 X num_species.}

\item{alphaij}{The interaction matrix of species num_species X num_species.}

\item{tmax}{The maximum time of the simulation.}

\item{mean_d}{The mean of the noise.}

\item{sd_d}{The standard deviation of the noise.}

\item{ht}{The number of substeps in the simulation.}
}
\value{
data frame of the densities of species.
}
\description{
Simulate a Lotka-Volterra model with noise.
}
\examples{
library(ggplot2)
library(reshape2)
# test
n0 <- c(1, 1, 1)
r <- c(1, 1, 1)
alpha <- matrix(c(0.2, 0.1, 0.3, 0.1, 0.5, 0.3, 0.1, 0.12, 0.3), nrow = 3, byrow = TRUE)
tmax <- 100
simresult <- simlv(n0 = n0, r = r, alphaij = alpha, tmax = tmax)

# long data frame
df <- melt(simresult, id.vars=c("time"))

# plot the result
ggplot(df, aes(x = time, y = value, group = variable, color = factor(variable))) + geom_line(linewidth = 2) + theme_minimal()
}
